refactor: extract firebase logic to FirebaseMatchRepository and optimize MatchService constructor arguments order
