+++ src/modules/match/match.repository.ts
@@ -1,4 +1,5 @@
 import { Match } from "../../models";
+import { firebaseDb } from "../../config/firebase";
 
 export interface IMatchRepository {
   create(user1_id: string, user2_id: string): Promise<any>;
@@ -10,6 +11,10 @@ export interface IMatchRepository {
     user2_id: string,
     update: Partial<{ user1_seen: boolean; user2_seen: boolean }>
   ): Promise<any | null>;
+}
+export interface IFirebaseRepository {
+  create(user1_id: string, user2_id: string): Promise<any>;
+  deleteMatch(user1_id: string, user2_id: string): Promise<any>;
 }
 
 export class MongoMatchRepository implements IMatchRepository {
@@ -56,5 +61,25 @@ export class MongoMatchRepository implements IMatchRepository {
         { user1_id: user2_id, user2_id: user1_id },
       ],
     }).exec();
+  }
+}
+
+export class FirebaseMatchRepository implements IFirebaseRepository {
+  async create(user1_id: string, user2_id: string): Promise<any> {
+    const comboId = [user1_id, user2_id].sort().join("_");
+    
+    await firebaseDb.ref("matches").child(comboId).set({
+      users: { [user1_id]: true, [user2_id]: true },
+      type: "match",
+      createdAt: new Date().toISOString()
+    });
+    
+    return { user1_id, user2_id, comboId };
+  }
+
+  async deleteMatch(user1_id: string, user2_id: string): Promise<any> {
+    const comboId = [user1_id, user2_id].sort().join("_");
+    await firebaseDb.ref("matches").child(comboId).remove();
+    return { success: true };
   }
 }
